Separate some of the new structs out into separate files.
